+++ shared/version-state/message.ts
@@ -1,0 +1,5 @@
+/** Single source of truth for the "welcome back" major-version-upgrade notice. */
+export const UPGRADE_NOTICE_MESSAGE =
+  "Hey! It looks like it's been a while since you've used this tool! There might be some changes " +
+  'you are not aware of, please read up on our changelog at ' +
+  'https://github.com/LiquidBuiltIt/Supersurf/blob/main/CHANGELOG.md';